requirements.txt
attrs==20.3.0
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
idna==2.10
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-engineio==4.0.1
python-socketio==5.1.0
requests==2.25.1
-e .
toml==0.10.2
urllib3==1.26.4
